k3s/batch/requirements.txt
mysql-connector-python==8.0.33
python-dotenv==1.0.0
urllib3==2.2.1
